chore(pagination): Move cursor based pagination to pkg
